requirements.txt
Django==2.2.7
django-bootstrap4
django-lti-provider